+++ backend/src/main/java/com/team15/sidedish/domain/BestRepository.java
@@ -5,12 +5,12 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface BestRepository extends CrudRepository<Best, Integer> {
+public interface BestRepository extends CrudRepository<Best, Long> {
 
     @Override
     List<Best> findAll();
 
     @Query("select dish_hash from dish_best where category_id = :categoryId")
-    List<String> findByCategoryId(Integer categoryId);
+    List<String> findByCategoryId(Long categoryId);
 
 }
